Move HelpPanel to wagtail.admin.panels.help_panel
